requirements.txt
aiocontextvars==0.2.2
contextvars==2.4
Django==2.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
immutables==0.15
install==1.3.4
opentelemetry-api==0.17b0
opentelemetry-exporter-jaeger==0.17b0
opentelemetry-instrumentation==0.17b0
opentelemetry-instrumentation-django==0.17b0
opentelemetry-instrumentation-wsgi==0.17b0
opentelemetry-sdk==0.17b0
protobuf==3.15.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
thrift==0.13.0
wrapt==1.12.1
